added stuff from pydata nyc 2013
