+++ prettyplotlib/_scatterhist.py
@@ -1,0 +1,17 @@
+__author__ = 'olga'
+
+
+def scatterhist(ax, x, y):
+    """
+    Create a scatterplot with histogram of the marginals, as in scatterhist()
+    in Matlab but nicer.
+
+    @param ax:
+    @type ax:
+    @param x:
+    @type x:
+    @param y:
+    @type y:
+    @return: hist_patches, scatter_points
+    @rtype: matplotlib.
+    """
